+++ CommonCode/CommonLibrary/getListKind.h
@@ -45,8 +45,8 @@
 # define A_ERROR 123 /* This should not be a legal t_atom type */
 
 # define GETLISTKIND_HEADER \
-unsigned char getListKind(const short argc,\
-                          t_atom *    argv)
+unsigned char getListKind(const long argc,\
+                          t_atom *   argv)
 
 GETLISTKIND_HEADER;
