+++ Vector Manipulation/Vdrop/Vdrop_In1.cpp
@@ -45,6 +45,6 @@ IN1_HEADER(VObjectData)
     /* We've received a number. */
     if (xx)
     {
-        xx->fHowMany = static_cast<short>(msg);
+        xx->fHowMany = msg;
     }
 } // cmd_In1
