+++ Vector Manipulation/Vexp/Vexp_List.cpp
@@ -59,7 +59,7 @@ LIST_HEADER(VObjectData)
                 t_atom * oldWalk = argv;
                 double   newValue;
 
-                for (short ii = 0; (ii < argc) && okSoFar; ++ii, ++newWalk, ++oldWalk)
+                for (long ii = 0; (ii < argc) && okSoFar; ++ii, ++newWalk, ++oldWalk)
                 {
                     newWalk->a_type = oldWalk->a_type;
                     switch (oldWalk->a_type)
